mahmoudhossam-CS253-wiki-9e64c74/requirements.txt
Django==1.4.1
argparse==1.2.1
distribute==0.6.24
dj-database-url==0.2.1
psycopg2==2.4.5
py-bcrypt==0.2
wsgiref==0.1.2
